requirements.txt
spacy
google-cloud-storage
git+https://github.com/explosion/spacy-transformers#spacy-transformers